Added configure-commands to enable DHCP and reboot device (hikvision)
The enable-dhcp command only work if we also reboot the camera
afterwards, which is only accomplished by sending another command, so we
need some way to note commands that depend on one-another. Maybe add
another field, "depends", and have it be a colon-separated list of
configure_commands to be run after the given command is ran?